fix log warning: A non-numeric value encountered in /srv/ocpl/lib/tm_ll_lib.php + viewcache coords disp. refactoring
