+++ lib/languages/cs.php
@@ -192,7 +192,7 @@ $translations = array(
     'user' => 'Uživatel',
     'password' => 'Heslo',
     'logged_as' => 'Přihlášeni jako',
-    'coords_other' => 'Souřadnice dalších systémů',
+    'viewCache_switchCoordsFormat' => 'Souřadnice dalších systémů',
     'size' => 'Velikost',
     'time' => 'Čas',
     'length' => 'Délka',
