+++ lib/languages/de.php
@@ -208,7 +208,7 @@ $translations = array(
     'user' => 'Benutzer',
     'password' => 'Passwort',
     'logged_as' => 'Eingeloggt als',
-    'coords_other' => 'Andere Koordinatensysteme',
+    'viewCache_switchCoordsFormat' => 'Andere Koordinatensysteme',
     'size' => 'Größe',
     'time' => 'Zeitaufwand',
     'length' => 'Wegstrecke',
