+++ lib/languages/en.php
@@ -208,7 +208,7 @@ $translations = array(
     'user' => 'User',
     'password' => 'Password',
     'logged_as' => 'Logged in as',
-    'coords_other' => 'Other coordinate systems',
+    'viewCache_switchCoordsFormat' => 'Change coordinates format',
     'size' => 'Size',
     'time' => 'Time required',
     'length' => 'Distance to travel',
