+++ lib/languages/es.php
@@ -191,7 +191,7 @@ $translations = array(
     'user' => 'Usuario',
     'password' => 'Contrasena',
     'logged_as' => 'Firmado en como',
-    'coords_other' => 'Las coordenadas de los otros sistemas',
+    'viewCache_switchCoordsFormat' => 'Las coordenadas de los otros sistemas',
     'size' => 'Tamano',
     'time' => 'Tiempo',
     'length' => 'Distancia',
