+++ lib/languages/fr.php
@@ -190,7 +190,7 @@ $translations = array(
     'user' => 'Utilisateur',
     'password' => 'Mot de passe',
     'logged_as' => 'Signé en tant que',
-    'coords_other' => 'Les coordonnées des autres systemes',
+    'viewCache_switchCoordsFormat' => 'Les coordonnées des autres systemes',
     'size' => 'Taille',
     'time' => 'Time',
     'length' => 'Distance',
