+++ lib/languages/nl.php
@@ -208,7 +208,7 @@ $translations = array(
     'user' => 'Gebruiker',
     'password' => 'Wachtwoord',
     'logged_as' => 'Ingelogd als',
-    'coords_other' => 'Andere coördinaatstelsels',
+    'viewCache_switchCoordsFormat' => 'Andere coördinaatstelsels',
     'size' => 'Grootte',
     'time' => 'Tijd',
     'length' => 'Lengte',
