+++ lib/languages/pl.php
@@ -208,7 +208,7 @@ $translations = array(
     'user' => 'Użytkownik',
     'password' => 'Hasło',
     'logged_as' => 'Zalogowany jako',
-    'coords_other' => 'Współrzędne w innych systemach',
+    'viewCache_switchCoordsFormat' => 'Zmień format współrzędnych',
     'size' => 'Wielkość',
     'time' => 'Czas',
     'length' => 'Długość trasy',
