+++ lib/languages/ro.php
@@ -208,7 +208,7 @@ $translations = array(
     'user' => 'Utilizator',
     'password' => 'Parolă',
     'logged_as' => 'Conectat ca',
-    'coords_other' => 'Alte sisteme de coordonate',
+    'viewCache_switchCoordsFormat' => 'Alte sisteme de coordonate',
     'size' => 'Mărime',
     'time' => 'Timp necesar',
     'length' => 'Distanţă de parcurs',
