+++ lib/languages/sv.php
@@ -189,7 +189,7 @@ $translations = array(
     'user' => 'Användare',
     'password' => 'Lösenord',
     'logged_as' => 'Inloggad som',
-    'coords_other' => 'Koordinater i andra system',
+    'viewCache_switchCoordsFormat' => 'Koordinater i andra system',
     'size' => 'Storlek',
     'time' => 'Tidsåtgång',
     'length' => 'Avstånd',
