--- tpl/stdstyle/coordinates.inc.php
@@ -1,3 +1,0 @@
-<?php
-
-?>
